textbook/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
qdrant-client==1.7.0
openai==1.10.0
pytest==7.4.4
httpx==0.26.0
